Add files via upload
In this project i did web scrapping using BeautifulSoup. First fetched the URL of IMDB 250 movies list. Then from that  separate the name of the movie, crew of the film ,rank of the film , ratings of the film and number of the users review and store in the the empty lists separately . Then from all that lists created a dataframe .